Refactoring - extracts the AddressUpdateForm as a separate component.
